#12: Physically splitting cxf into two discrete 'modules'
