+++ common/cxf-server/BUCK
@@ -1,0 +1,19 @@
+java_library(
+    name = "queue-triage-common-cxf-server",
+    srcs = glob(
+        ["src/main/java/**/*.java"],
+    ),
+    exported_deps = [
+        "//common/cxf:queue-triage-common-cxf",
+        "//lib/apache-cxf:cxf-rt-transports-http-jetty",
+        "//lib/javax:javax-servlet-api",
+        "//lib/spring:spring-boot",
+        "//lib/spring:spring-context",
+        "//lib/spring:spring-core",
+    ],
+    visibility = [
+        "//core/component-test:test",
+        "//core/server:queue-triage-core-server",
+        "//web/server:queue-triage-web-server",
+    ],
+)
