+++ common/cxf/src/main/java/uk/gov/dwp/queue/triage/cxf/CxfConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.dwp.migration.mongo.demo.cxf.client;
+package uk.gov.dwp.queue.triage.cxf;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
