+++ common/jackson/BUCK
@@ -2,7 +2,7 @@ java_library(
     name = "queue-triage-common-jackson",
     srcs = glob(["src/main/java/**/*.java"], excludes = ["src/main/java/**/configuration/*.java"]),
     exported_deps =[
-        "//common/cxf:queue-triage-cxf-client",
+        "//common/cxf:queue-triage-common-cxf",
         "//common/id:queue-triage-common-id",
         "//lib/javax:javax-ws-rs-api",
         '//lib/jackson:jackson-datatype-jdk8',
