+++ common/swagger/BUCK
@@ -2,7 +2,7 @@ java_library(
     name = 'queue-triage-common-swagger-cxf',
     srcs = glob(['src/main/java/**/*.java']),
     exported_deps = [
-        '//common/cxf:queue-triage-cxf-client',
+        '//common/cxf:queue-triage-common-cxf',
         '//lib/apache-cxf:cxf-core',
         '//lib/apache-cxf:cxf-rt-rs-service-description-swagger',
         '//lib/spring:spring-context',
