+++ core/server/BUCK
@@ -11,8 +11,8 @@ java_library(
         "//lib/spring:spring-jms",
         "//lib/spring:spring-tx",
         "//lib/spring:spring-web",
-        '//common/cxf:queue-triage-cxf',
-        '//common/cxf:queue-triage-cxf-client',
+        '//common/cxf:queue-triage-common-cxf',
+        '//common/cxf-server:queue-triage-common-cxf-server',
         '//common/health:queue-triage-health',
         '//common/id:queue-triage-common-id',
         '//common/jackson:queue-triage-common-jackson-spring',
