+++ web/mustache/src/main/java/uk/gov/dwp/queue/triage/web/common/mustache/configuration/MustacheConfiguration.java
@@ -5,8 +5,8 @@ import com.github.mustachejava.MustacheFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import uk.gov.dwp.migration.mongo.demo.cxf.client.CxfConfiguration;
-import uk.gov.dwp.migration.mongo.demo.cxf.client.ProviderRegistry;
+import uk.gov.dwp.queue.triage.cxf.CxfConfiguration;
+import uk.gov.dwp.queue.triage.cxf.ProviderRegistry;
 import uk.gov.dwp.queue.triage.web.common.mustache.MustachePageRenderer;
 import uk.gov.dwp.queue.triage.web.common.mustache.PageMessageBodyWriter;
